+++ src/components/ui/Dropdown/handlers.ts
@@ -2,6 +2,7 @@ import type { DropdownTheme } from '../../../types/theme';
 import type { Size } from '../../../types/sizes';
 import { getDropdownItemStyles } from '../../../handlers/dropdownThemeHandlers';
 import type React from 'react';
+import type { NullableRefObject } from '../../../types/reactRefs';
 import {
   HintPosition,
   TooltipPosition,
@@ -430,8 +431,8 @@ export const getMenuItemKey = (
  */
 export const isClickInsideDropdown = (
   event: MouseEvent,
-  triggerRef: React.RefObject<HTMLElement>,
-  menuRef: React.RefObject<HTMLElement>,
+  triggerRef: NullableRefObject,
+  menuRef: NullableRefObject,
 ): boolean => {
   const target = event.target as Node;
   const isClickInsideTrigger = triggerRef.current && triggerRef.current.contains(target);
@@ -450,8 +451,8 @@ export const isClickInsideDropdown = (
  */
 export const handleClickOutsideEvent = (
   event: MouseEvent,
-  triggerRef: React.RefObject<HTMLElement>,
-  menuRef: React.RefObject<HTMLElement>,
+  triggerRef: NullableRefObject,
+  menuRef: NullableRefObject,
   onClickOutside?: (event: Event) => void,
 ): boolean => {
   const isOutside = !isClickInsideDropdown(event, triggerRef, menuRef);